requirements.txt
# Fetch.AI / uAgents
uagents>=0.1.0
fetchai>=0.1.0

# HTTP client
httpx>=0.27.0

# LLM - Groq
groq>=0.4.0

# Image Generation - Google Vertex AI Imagen
google-cloud-aiplatform>=1.38.0

# Image Generation - Replicate (backup)
replicate>=0.25.0

# Text-to-Speech - ElevenLabs
elevenlabs>=1.0.0

# HTTP Requests
requests>=2.31.0
aiohttp>=3.9.0

# PDF Generation
reportlab>=4.0.0
Pillow>=10.0.0

# Environment & Config
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# OpenAI-compatible client (for ASI:One if needed)
openai>=1.0.0
